+++ modules/suggestions.py
@@ -1,7 +1,6 @@
 from modules.api_request import bgg_api_call
-from modules.db import run_query
-from modules.designers import get_games_from_designer
-from modules.boardgame import save_games
+from modules.db import save_list_network_to_db
+from config.db_connection import run_query
 import pandas as pd
 import numpy as np
 import ast
@@ -11,7 +10,8 @@ def suggest_games(
     user, game_status={"own": 1, "stats": 1}, source="rating", results=5, top=5
 ):
     user_collection = bgg_api_call("collection", user, game_status)
-    games_from_user_network(user_collection)
+    games_id = [x["@objectid"] for x in user_collection if x["@subtype"] == "boardgame"]
+    save_list_network_to_db(games_id)
     df_plays = (
         pd.DataFrame(
             [
@@ -115,35 +115,3 @@ def get_mechanics_best(list_game_id, user_collection, no_expansion=True, results
         ["name", "rating", "type"]
     ].reset_index(drop=True)
     return df_mechanics_top.assign(recommendation="mechanics")
-
-
-def games_from_user_network(user_collection):
-    games = [x["@objectid"] for x in user_collection if x["@subtype"] == "boardgame"]
-    save_games(games)
-
-    query = f"""select distinct designer
-                from boardgame
-                where id in ({','.join(games)})"""
-    df_designers = run_query(query, execute_only=False)
-
-    df_designer_game_count = run_query(
-        "select designer,count(*) as games_count from boardgame b group by 1",
-        execute_only=False,
-    )
-    df_designer_download = df_designers.merge(
-        df_designer_game_count, how="inner", on="designer"
-    )
-    df_designer_download = df_designer_download[df_designer_download.games_count == 1]
-    count_designer = len(df_designer_download)
-    print("###################################################")
-    print("Processing", count_designer, "designers")
-    print("###################################################")
-    i = 0
-    for designer in df_designer_download.to_dict("records"):
-        i += 1
-        print("###################################################")
-        print(
-            f"Designer {str(i).zfill(2)}/{str(count_designer).zfill(2)}: Processing {designer['designer']}"
-        )
-        print("###################################################")
-        get_games_from_designer(designer["designer"])
